Refactoring, separating components
